goes up to 88% win rate against random

+++ Four in a row/BotBase.cs
@@ -9,7 +9,7 @@ namespace Learning
 
     public abstract class LearningBot : Bot
     {
-        public abstract void Learn(int EpocheNumber, float EpsilonLimit, float EpsilonDecrease, float LearningRate, float DiscountRate, Bot against = null);
+        public abstract void Learn(int EpocheNumber, double EpsilonLimit, double EpsilonDecrease, double LearningRate, double DiscountRate, Bot against = null);
         public abstract void Stop();
         public abstract void Setup(GameControlBase control, GameControlBase.Players player);
         public GameControlBase.Players BotTurn { get; protected set; }
